requirements.txt
blinker==1.9.0
click==8.3.0
colorama==0.4.6
dnspython==2.8.0
email-validator==2.3.0
Flask==3.1.2
Flask-Login==0.6.3
Flask-PyMongo==3.0.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.4
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mongomock==4.3.0
numpy==2.3.4
packaging==25.0
pandas==2.3.3
pluggy==1.6.0
Pygments==2.19.2
pymongo==4.15.3
pytest==9.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
sentinels==1.1.1
six==1.17.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
tzdata==2025.2
Werkzeug==3.1.3
WTForms==3.2.1
